# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.

package db

import (
	"database/sql"
)

type Arrival struct {
	ID        int32 `json:"id"`
	GuestID   int32 `json:"guest_id"`
	TableID   int32 `json:"table_id"`
	PartySize int32 `json:"party_size"`
}

type Guest struct {
	ID          int32        `json:"id"`
	GuestName   string       `json:"guest_name"`
	Entourage   int32        `json:"entourage"`
	TableID     int32        `json:"table_id"`
	ArrivalTime sql.NullTime `json:"arrival_time"`
	CreatedAt   sql.NullTime `json:"created_at"`
}

type Table struct {
	ID        int32        `json:"id"`
	Size      int32        `json:"size"`
	Occupied  int32        `json:"occupied"`
	CreatedAt sql.NullTime `json:"created_at"`
}